+++ src/models/ReservationDate.js
@@ -1,8 +1,9 @@
 import { DATE_REGEX_PATTERN } from '../constants/RegexPattern.js';
+import { ErrorController } from '../controllers/index.js';
 import { getReservationDateErrorMessage, testRegExp } from '../utils/index.js';
-import CustomError from '../customError/index.js';
 
 class ReservationDate {
+  #date;
   constructor(string) {
     this.#validateDate(string);
     this.#setDate(string);
@@ -13,11 +14,14 @@ class ReservationDate {
   #validateDate(string) {
     const message = getReservationDateErrorMessage();
     if (!this.#isSuitableForDate(string))
-      new CustomError('reservation date error', message);
+      new ErrorController('reservation date error', message).throwError();
   }
   #setDate(string) {
     const date = Number(string);
-    return date;
+    this.#date = date;
+  }
+  getDate() {
+    return this.#date;
   }
 }
